modified file structure to separate react and redux parts and display users favorite spots when clicked
